survey_api/requirements.txt
asgiref==3.8.1
astroid==2.15.8
attrs==24.3.0
autobahn==24.4.2
autoflake==2.3.1
Automat==24.8.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docformatter==1.7.5
drf-yasg==1.21.8
et_xmlfile==2.0.0
flake8==6.1.0
flake8-bugbear==24.12.12
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==2.2.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.1.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
snowballstemmer==2.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
untokenize==0.1.1
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
zope.interface==7.2
